+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/helper/WireMockInitializer.java
@@ -1,0 +1,39 @@
+package onlydust.com.marketplace.api.bootstrap.helper;
+
+import lombok.NonNull;
+import org.springframework.context.ApplicationContextInitializer;
+import org.springframework.context.ConfigurableApplicationContext;
+
+public class WireMockInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+    @Override
+    public void initialize(final @NonNull ConfigurableApplicationContext context) {
+        WiremockServerRegistration.builder()
+                .beanName("ethereumWireMockServer")
+                .stubLocation("ethereum")
+                .property("infrastructure.ethereum.base-uri")
+//                    .recordFrom("https://mainnet.infura.io/v3")
+                .build()
+                .register(context);
+
+        WiremockServerRegistration.builder()
+                .beanName("optimismWireMockServer")
+                .stubLocation("optimism")
+                .property("infrastructure.optimism.base-uri")
+                .build()
+                .register(context);
+
+        WiremockServerRegistration.builder()
+                .beanName("starknetWireMockServer")
+                .stubLocation("starknet")
+                .property("infrastructure.starknet.base-uri")
+                .build()
+                .register(context);
+
+        WiremockServerRegistration.builder()
+                .beanName("coinmarketcapWireMockServer")
+                .stubLocation("coinmarketcap")
+                .property("infrastructure.coinmarketcap.base-uri")
+                .build()
+                .register(context);
+    }
+}
